Add net.sf.ehcache:ehcache:2.8.3
The Ehcache team have indicated that net.sf.ehcache:ehcache-core
should not be used (this is made clear by more recent versions having
a group id of net.sf.ehcache.internal) and that net.sf.ehcache:ehcache
is the preferred dependency.

This commit adds dependency management for net.sf.ehcache:ehcache.
To preserve backwards compatibility, the net.sf.ehcache:ehcache-core
remains but it will be documented as being deprecated and will be
removed in a future release of the Platform.

spring-io/platform#28