Use a more sophisticated approach for refreshing Gradle dependencies
Rather than using --refresh-dependencies, which refreshes every
dependency, use a resolution strategy to turn off caching for
changing dependencies. When using Maven repositories, which we are,
Gradle considers all -SNAPSHOT dependencies to be changing so this
is sufficient for our needs and avoid the cost of unnecessarily
refresh dependencies that we know will not have changed.