Filter out Jetty's WebSocket dependencies where possible
Platform 1.1 uses Jetty 8 and does so by overriding Boot's Jetty
version property. This has the unwanted side-effect of leaving
dependency management in place for Jetty's WebSocket modules using
a version of Jetty that isn't supported by Spring's WebSocket support
and where some of the artifacts don't exist.

Unfortunately, it isn't possible to remove this dependency management
from the Platform's bom, but we can remove it from the properties
files and the documentation. Where it remains in the Maven bom it
should prove to be harmless for Jetty 8 users and will be of use to
those who override the Jetty version to 9.x

See spring-io/platform#103